Testing update and updated according to PR comments
