+++ package.json
@@ -96,7 +96,7 @@
     "@testing-library/user-event": "^14.5.2",
     "babel-jest": "^29.7.0",
     "gh-pages": "^6.1.1",
-    "jest": "^28.0.0",
+    "jest": "^28.1.3",
     "jest-environment-jsdom": "^29.7.0",
     "jest-watch-typeahead": "^2.2.2",
     "react-test-renderer": "^18.2.0"
