+++ scripts/build.js
@@ -69,24 +69,23 @@ checkBrowsers(paths.appPath, isInteractive)
   })
   .then(
     ({ stats, previousFileSizes, warnings }) => {
-      if (warnings.length) {
-        console.log(chalk.yellow('Compiled with warnings.\n'));
-        console.log(warnings.join('\n\n'));
-        console.log(
-          '\nSearch for the ' +
-            chalk.underline(chalk.yellow('keywords')) +
-            ' to learn more about each warning.'
-        );
-        console.log(
-          'To ignore, add ' +
-            chalk.cyan('// eslint-disable-next-line') +
-            ' to the line before.\n'
-        );
-      } else {
-        console.log(chalk.green('Compiled successfully.\n'));
-      }
+      // if (warnings.length) {
+      //   console.log(chalk.yellow('Compiled with warnings.\n'));
+      //   console.log(warnings.join('\n\n'));
+      //   console.log(
+      //     '\nSearch for the ' +
+      //       chalk.underline(chalk.yellow('keywords')) +
+      //       ' to learn more about each warning.'
+      //   );
+      //   console.log(
+      //     'To ignore, add ' +
+      //       chalk.cyan('// eslint-disable-next-line') +
+      //       ' to the line before.\n'
+      //   );
+      // } else {
+      //   console.log(chalk.green('Compiled successfully.\n'));
+      // }
 
-      console.log('File sizes after gzip:\n');
       printFileSizesAfterBuild(
         stats,
         previousFileSizes,
@@ -94,7 +93,6 @@ checkBrowsers(paths.appPath, isInteractive)
         WARN_AFTER_BUNDLE_GZIP_SIZE,
         WARN_AFTER_CHUNK_GZIP_SIZE
       );
-      console.log();
 
       const appPackage = require(paths.appPackageJson);
       const publicUrl = paths.publicUrlOrPath;
@@ -133,7 +131,6 @@ checkBrowsers(paths.appPath, isInteractive)
 
 // Create the production build and print the deployment instructions.
 function build(previousFileSizes) {
-  console.log('Creating an optimized production build...');
 
   const compiler = webpack(config);
   return new Promise((resolve, reject) => {
