+++ scripts/start.js
@@ -56,12 +56,8 @@ if (process.env.HOST) {
     )
   );
   console.log(
-    `If this was unintentional, check that you haven't mistakenly set it in your shell.`
-  );
-  console.log(
     `Learn more here: ${chalk.yellow('https://cra.link/advanced-config')}`
   );
-  console.log();
 }
 
 // We require that you explicitly set browsers and do not fall back to
@@ -127,7 +123,6 @@ checkBrowsers(paths.appPath, isInteractive)
         );
       }
 
-      console.log(chalk.cyan('Starting the development server...\n'));
       openBrowser(urls.localUrlForBrowser);
     });
